Fix apphost permission issues with .NET Core 3.
.NET Core 3 introduces a number of problems for nix/nixos. The main
problem is that dotnet wants copy, and locally modify, the "apphost" binary into
the obj directory. If apphost is copied with read-only permissions from
the store, the build fails since the binary cannot be modified. The
solution is to wrap the dotnet binary with a script which points
DOTNET_ROOT to a local sdk in $HOME/.dotnet/sdk/${version}, which
contains a rw copy of apphost.